+++ libraries/exoplayer_ima/src/main/java/androidx/media3/exoplayer/ima/ImaUtil.java
@@ -40,20 +40,18 @@ import androidx.media3.common.AdPlaybackState;
 import androidx.media3.common.AdPlaybackState.AdGroup;
 import androidx.media3.common.AdViewProvider;
 import androidx.media3.common.C;
-import androidx.media3.common.Player;
 import androidx.media3.common.Timeline;
 import androidx.media3.common.util.Util;
 import androidx.media3.datasource.DataSchemeDataSource;
 import androidx.media3.datasource.DataSourceUtil;
 import androidx.media3.datasource.DataSpec;
-import com.google.ads.interactivemedia.v3.api.Ad;
 import com.google.ads.interactivemedia.v3.api.AdDisplayContainer;
 import com.google.ads.interactivemedia.v3.api.AdError;
 import com.google.ads.interactivemedia.v3.api.AdErrorEvent;
 import com.google.ads.interactivemedia.v3.api.AdEvent;
 import com.google.ads.interactivemedia.v3.api.AdPodInfo;
+import com.google.ads.interactivemedia.v3.api.AdSlot;
 import com.google.ads.interactivemedia.v3.api.AdsLoader;
-import com.google.ads.interactivemedia.v3.api.AdsManager;
 import com.google.ads.interactivemedia.v3.api.AdsRenderingSettings;
 import com.google.ads.interactivemedia.v3.api.AdsRequest;
 import com.google.ads.interactivemedia.v3.api.CompanionAdSlot;
@@ -181,6 +179,7 @@ import java.util.Set;
     @Nullable public final AdEvent.AdEventListener applicationAdEventListener;
     @Nullable public final AdErrorEvent.AdErrorListener applicationAdErrorListener;
     public final ImmutableList<CompanionAdSlot> companionAdSlots;
+    @Nullable public final AdSlot pauseAdSlot;
     public final boolean focusSkipButtonWhenAvailable;
     public final boolean enableCustomTabs;
     public final boolean debugModeEnabled;
@@ -192,6 +191,7 @@ import java.util.Set;
         @Nullable AdEvent.AdEventListener applicationAdEventListener,
         @Nullable AdErrorEvent.AdErrorListener applicationAdErrorListener,
         List<CompanionAdSlot> companionAdSlots,
+        @Nullable AdSlot pauseAdSlot,
         boolean focusSkipButtonWhenAvailable,
         boolean enableCustomTabs,
         boolean debugModeEnabled) {
@@ -201,6 +201,7 @@ import java.util.Set;
       this.applicationAdEventListener = applicationAdEventListener;
       this.applicationAdErrorListener = applicationAdErrorListener;
       this.companionAdSlots = ImmutableList.copyOf(companionAdSlots);
+      this.pauseAdSlot = pauseAdSlot;
       this.focusSkipButtonWhenAvailable = focusSkipButtonWhenAvailable;
       this.enableCustomTabs = enableCustomTabs;
       this.debugModeEnabled = debugModeEnabled;